+++ AddOns/IcyVeinsStatPriority/Data.lua
@@ -1,8 +1,6 @@
 local addonName, IVSP = ...
 local L = IVSP.L
 
--- ≥
-
 local data = {
     -- 250 - Death Knight: Blood -- https://www.icy-veins.com/wow/blood-death-knight-pve-tank-stat-priority
     [250] = {
@@ -20,8 +18,8 @@ local data = {
 
     -- 577 - Demon Hunter: Havoc -- https://www.icy-veins.com/wow/havoc-demon-hunter-pve-dps-stat-priority
     [577] = {
-        {"Agility > Critical Strike > Mastery > Haste > Versatility", "單目標"},
-        {"Agility > Critical Strike > Mastery > Versatility > Haste", "多目標"},
+        {"Mastery = Critical Strike > Agility > Haste > Versatility", "Fel-Scarred"},
+        {"Critical Strike = Agility > Mastery > Versatility > Haste", "Aldrachi"},
     },
     -- 581 - Demon Hunter: Vengeance -- https://www.icy-veins.com/wow/vengeance-demon-hunter-pve-tank-stat-priority
     [581] = {
@@ -30,13 +28,13 @@ local data = {
 
     -- 102 - Druid: Balance -- https://www.icy-veins.com/wow/balance-druid-pve-dps-stat-priority
     [102] = {
-        {"Intellect > Mastery > Versatility > Haste > Critical Strike"},
+        {"Intellect > Mastery > Haste > Versatility > Critical Strike"},
     },
     -- 103 - Druid: Feral -- https://www.icy-veins.com/wow/feral-druid-pve-dps-stat-priority
     [103] = {
         {"Critical Strike = Mastery > Agility > Haste > Versatility", "單目標"},
         {"Mastery > Critical Strike = Haste > Agility > Versatility", "多目標 (野地潛獵者)"},
-        {"Mastery > Agility > Critical Strike > Versatility = Haste", "多目標 (利爪)"},
+        {"Mastery > Agility > Critical Strike > Haste > Versatility", "多目標 (利爪)"},
     },
     -- 104 - Druid: Guardian -- https://www.icy-veins.com/wow/guardian-druid-pve-tank-stat-priority
     [104] = {
@@ -56,40 +54,42 @@ local data = {
     },
     -- Evoker: Preservation -- https://www.icy-veins.com/wow/preservation-evoker-pve-healing-stat-priority
     [1468] = {
-        {"Intellect > Mastery > Critical Strike > Versatility > Haste", "團隊"},
+        {"Intellect > Mastery > Critical Strike > Haste > Versatility", "團隊"},
         {"Intellect > Critical Strike > Haste > Versatility > Mastery", "M+"},
     },
     -- Evoker -- https://www.icy-veins.com/wow/augmentation-evoker-pve-dps-stat-priority
     [1473] = {
         {"Intellect > Haste (10%) > Mastery = Critical Strike > Haste > Versatility", "時光看守者"},
-        {"Intellect > Haste (10%) > Mastery > Critical Strike = Haste > Versatility", "龍隊指揮官"},
+        {"Intellect > Haste (20%) > Mastery > Critical Strike = Haste > Versatility", "龍隊指揮官"},
     },
 
     -- 253 - Hunter: Beast Mastery -- https://www.icy-veins.com/wow/beast-mastery-hunter-pve-dps-stat-priority
     [253] = {
-        {"Haste > Critical Strike > Mastery > Versatility", "單目標"},
-        {"Mastery > Haste > Critical Strike > Versatility", "多目標"},
+        {"Critical Strike > Haste > Mastery > Versatility", "獸群領袖"},
+        {"Haste > Critical Strike > Mastery > Versatility", "黑暗遊俠"},
     },
     -- 254 - Hunter: Marksmanship -- https://www.icy-veins.com/wow/marksmanship-hunter-pve-dps-stat-priority
     [254] = {
-        {"Weapon Damage > Critical Strike > Mastery >  > Versatility > Haste"},
+        {"Weapon Damage > Critical Strike > Mastery > Agility > Versatility > Haste"},
     },
     -- 255 - Hunter: Survival -- https://www.icy-veins.com/wow/survival-hunter-pve-dps-stat-priority
     [255] = {
-        {"Mastery > Agility > Haste > Critical Strike > Versatility"},
+        {"Agility > Critical Strike > Mastery > Haste > Versatility", "獸群領袖 (單目標)"},
+        {"Agility > Critical Strike > Haste = Mastery > Versatility", "獸群領袖 (多目標)"},
+        {"Agility > Mastery > Haste = Critical Strike > Versatility", "Sentinel"},
     },
 
     -- 62 - Mage: Arcane -- https://www.icy-veins.com/wow/arcane-mage-pve-dps-stat-priority
     [62] = {
-        {"Intellect > Mastery = Haste ≥ Versatility ≥ Critical Strike"},
+        {"Intellect > Haste > Mastery > Versatility > Critical Strike"},
     },
     -- 63 - Mage: Fire -- https://www.icy-veins.com/wow/fire-mage-pve-dps-stat-priority
     [63] = {
-        {"Intellect > Haste > Versatility > Mastery > Critical Strike"},
+        {"Intellect > Haste >> Versatility > Mastery > Critical Strike"},
     },
     -- 64 - Mage: Frost -- https://www.icy-veins.com/wow/frost-mage-pve-dps-stat-priority
     [64] = {
-        {"Intellect > Mastery > Haste > Critical Strike (33.34%) > Versatility"},
+        {"Intellect > Haste > Critical Strike (33.34%) > Versatility > Mastery"},
     },
 
     -- 268 - Monk: Brewmaster -- https://www.icy-veins.com/wow/brewmaster-monk-pve-tank-stat-priority
@@ -109,8 +109,8 @@ local data = {
 
     -- 65 - Paladin: Holy -- https://www.icy-veins.com/wow/holy-paladin-pve-healing-stat-priority
     [65] = {
-        {"Intellect > Critical Strike > Haste > Mastery > Versatility", "團隊"},
-        {"Intellect > Critical Strike > Haste > Versatility > Mastery", "M+"},
+        {"Intellect > Haste > Mastery > Versatility", "團隊"},
+        {"Intellect > Haste > Versatility > Critical Strike > Mastery", "M+"},
     },
     -- 66 - Paladin: Protection -- https://www.icy-veins.com/wow/protection-paladin-pve-tank-stat-priority
     [66] = {
@@ -118,7 +118,7 @@ local data = {
     },
     -- 70 - Paladin: Retribution -- https://www.icy-veins.com/wow/retribution-paladin-pve-dps-stat-priority
     [70] = {
-        {"Mastery > Strength > Critical Strike = Haste > Versatility"},
+        {"Strength > Mastery > Critical Strike = Haste > Versatility"},
     },
 
     -- 256 - Priest: Discipline -- https://www.icy-veins.com/wow/discipline-priest-pve-healing-stat-priority
@@ -150,17 +150,16 @@ local data = {
 
     -- 262 - Shaman: Elemental -- https://www.icy-veins.com/wow/elemental-shaman-pve-dps-stat-priority
     [262] = {
-        {"Intellect > Haste >> Critical Strike > Versatility >> Mastery", "電"},
-        {"Intellect > Haste >> Versatility > Mastery >> Critical Strike", "火"},
+        {"Intellect > Mastery > Haste > Critical Strike > Versatility "},
     },
     -- 263 - Shaman: Enhancement -- https://www.icy-veins.com/wow/enhancement-shaman-pve-dps-stat-priority
     [263] = {
         {"Haste = Agility > Mastery > Critical Strike > Versatility", "風暴使者"},
-        {"Agility = Haste > Mastery > Versatility > Critical Strike", "圖騰使者"},
+        {"Haste = Mastery > Agility > Versatility > Critical Strike", "圖騰使者"},
     },
     -- 264 - Shaman: Restoration -- https://www.icy-veins.com/wow/restoration-shaman-pve-healing-stat-priority
     [264] = {
-        {"Intellect > Versatility = Critical Strike > Haste = Mastery"},
+        {"Intellect > Critical Strike > Versatility > Haste = Mastery"},
     },
 
     -- 265 - Warlock: Affliction -- https://www.icy-veins.com/wow/affliction-warlock-pve-dps-stat-priority