+++ AddOns/IcyVeinsStatPriority/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Icy Veins Stat Priority
 
-## [2024-08-24](https://github.com/enderneko/IcyVeinsStatPriority/tree/2024-08-24) (2024-08-24)
-[Full Changelog](https://github.com/enderneko/IcyVeinsStatPriority/compare/2024-05-09...2024-08-24) [Previous Releases](https://github.com/enderneko/IcyVeinsStatPriority/releases)
+## [2025-01-02](https://github.com/enderneko/IcyVeinsStatPriority/tree/2025-01-02) (2025-01-04)
+[Full Changelog](https://github.com/enderneko/IcyVeinsStatPriority/compare/2024-08-24...2025-01-02) [Previous Releases](https://github.com/enderneko/IcyVeinsStatPriority/releases)
 
-- Update for tww  
+- Thanks Arseynimz for updating all the data  
+- Update frame level  
+- Update FUNDING.yml